+++ NEWS.md
@@ -1,2 +1,7 @@
+# bs4Dash 0.2.0.9000
+- update `bs4DashGallery()`. Add a "theme" argument
+- fix vignette titles
+- update Readme
+
 # bs4Dash 0.1.0
 - initial release